requirements_dev.txt
cx_Freeze==8.4.0
markdown2==2.5.4
pillow==11.3.0
polib==1.2.0
pytest==8.4.1
pymupdf==1.26.3
pywin32==311
requests==2.32.5
tkinterweb==4.4.4
coverage
pytest-cov